requirements.txt
fastapi
jinja2
uvicorn
python-dotenv
sqlalchemy
pymysql
python-multipart
pyopenssl